Rebase on top of main + clean up parent db class and events db
