--- overlore/sqlite/queries.py
@@ -1,2 +1,0 @@
-INSERT_QUERY = """"""
-MOST_RELEVANT_QUERY = """"""
