+++ overlore/sqlite/utils.py
